7 small miscellaneous fixes
	Bug#959 - Multiple minor improvements

	* configure.ac: pick up GDKPIXBUF_PREFIX from gdk-pixbuf-2.0
	module, not pango

	* clutter/clutter-keysyms-table.h: cosmetic fix for ifdef guard
	name in the comment (leftover from earlier commit)

	* tests/test-textures.c: don't depend on GdkPixbuf as we're not
	using any actual features from it, just plain pixel buffer
	manipulation

	* clutter/clutter-timeline.c
	* clutter/clutter-timeline.h (clutter_timeline_list_markers): Use
	gsize* to return number of items, not guint*
	* tests/test-paint-wrapper.c: use correct type (guint*) in call
	to clutter_actor_get_size()

	* tests/test-depth.c (janus_group): properly take height2 into
	account when calculating needed height for the rectangle

	* tests/test-cogl-tex-getset.c: use rowstride, not width*4, when
	calculating pixel offsets

	* tests/test-cogl-tex-getset.c: Don't assume/force RGBA format,
	also support ARGB format (needed with quartz imageloader.)

+++ clutter/clutter-timeline.h
@@ -153,7 +153,7 @@ void             clutter_timeline_remove_marker         (ClutterTimeline *timeli
                                                          const gchar     *marker_name);
 gchar **         clutter_timeline_list_markers          (ClutterTimeline *timeline,
                                                          gint             frame_num,
-                                                         guint           *n_markers) G_GNUC_MALLOC;
+                                                         gsize           *n_markers) G_GNUC_MALLOC;
 gboolean         clutter_timeline_has_marker            (ClutterTimeline *timeline,
                                                          const gchar     *marker_name);
 void             clutter_timeline_advance_to_marker     (ClutterTimeline *timeline,
